Extracted shared logic to helper
